+++ examples/autoscaling/main.tf
@@ -97,7 +97,7 @@ module "dynamodb_table" {
     ]
   }
 
-  attributes = [
+  table_attributes = [
     {
       name = "id"
       type = "N"
@@ -132,6 +132,8 @@ module "dynamodb_table" {
 
 module "disabled_dynamodb_table" {
   source = "../../"
+
+  table_attributes = []
 
   create_table = false
 }