+++ examples/global-tables/main.tf
@@ -50,7 +50,7 @@ module "dynamodb_table" {
   server_side_encryption_enabled     = true
   server_side_encryption_kms_key_arn = aws_kms_key.primary.arn
 
-  attributes = [
+  table_attributes = [
     {
       name = "id"
       type = "N"