+++ schedulescaling.tf
@@ -1,5 +1,5 @@
 resource "aws_appautoscaling_scheduled_action" "table_read_schedule" {
-  for_each = var.create_table && var.autoscaling_enabled && length(var.schedule_scaling_read) > 0 ? { for k, v in var.schedule_scaling_read : k => v } : {}
+  for_each = module.this.enabled && var.create_table && var.autoscaling_enabled && length(var.schedule_scaling_read) > 0 ? { for k, v in var.schedule_scaling_read : k => v } : {}
 
   name = "DynamoDBReadCapacityUtilization-${replace(aws_appautoscaling_target.table_read[0].resource_id, "/", "-")}-${each.key}"
 
@@ -16,7 +16,7 @@ resource "aws_appautoscaling_scheduled_action" "table_read_schedule" {
 }
 
 module "index_read_schedule" {
-  for_each = var.create_table && length(var.schedule_scaling_indexes_read) > 0 ? var.autoscaling_indexes : {}
+  for_each = module.this.enabled && var.create_table && length(var.schedule_scaling_indexes_read) > 0 ? var.autoscaling_indexes : {}
   source   = "./module"
 
   name = "DynamoDBReadCapacityUtilization-${replace(aws_appautoscaling_target.index_read[each.key].resource_id, "/", "-")}"
@@ -29,7 +29,7 @@ module "index_read_schedule" {
 }
 
 resource "aws_appautoscaling_scheduled_action" "table_write_schedule" {
-  for_each = var.create_table && var.autoscaling_enabled && length(var.schedule_scaling_write) > 0 ? { for k, v in var.schedule_scaling_write : k => v } : {}
+  for_each = module.this.enabled && var.create_table && var.autoscaling_enabled && length(var.schedule_scaling_write) > 0 ? { for k, v in var.schedule_scaling_write : k => v } : {}
 
   name = "DynamoDBWriteCapacityUtilization-${replace(aws_appautoscaling_target.table_write[0].resource_id, "/", "-")}-${each.key}"
 
@@ -46,7 +46,7 @@ resource "aws_appautoscaling_scheduled_action" "table_write_schedule" {
 }
 
 module "index_write_schedule" {
-  for_each = var.create_table && length(var.schedule_scaling_indexes_write) > 0 ? var.autoscaling_indexes : {}
+  for_each = module.this.enabled && var.create_table && length(var.schedule_scaling_indexes_write) > 0 ? var.autoscaling_indexes : {}
   source   = "./module"
 
   name = "DynamoDBWriteCapacityUtilization-${replace(aws_appautoscaling_target.index_write[each.key].resource_id, "/", "-")}"